whisper_to_s3/requirements.txt
git+https://github.com/openai/whisper.git
boto3
pydub
python-dotenv
ffmpeg-python
kafka-python